fix: return int brightness; return HS 2-tuple from get_color for HA hs_color
Brightness calculation returned a float (127.5 for 50%), but HA light
entities require int. get_color returned an RGB 3-tuple (r,g,b) but
HA hs_color expects (hue_degrees, saturation_percent) — Hive API stores
these natively so no conversion is needed.

+++ src/devices/light.py
@@ -62,7 +62,7 @@ class HiveLight(LightColorHandler, BaseDeviceHandler):
         try:
             data = self.session.data.products[device.hive_id]
             state = data["state"]["brightness"]
-            final = (state / 100) * 255
+            final = int((state / 100) * 255)
         except KeyError as e:
             _LOGGER.error(
                 "KeyError getting light brightness for %s: %s", device_name, str(e)
